requirements.txt
fastapi~=0.115.8
uvicorn==0.15.0
sqlalchemy==1.4.23
alembic~=1.12.0
PyYAML~=6.0.2
sqlite-utils~=3.15.0
pydantic~=2.10.6
injector~=0.22.0